backend/requirements.txt
fastapi
uvicorn
pydantic
requests
pillow
numpy
easyocr
pdf2image
crawl4ai
crewai
PyPDF2
python-multipart